+++ src/components/DatePicker/index.ts
@@ -1,2 +1,7 @@
 export { DatePicker } from './DatePicker';
-export type { DatePickerProps } from './DatePicker';
+export type {
+  DatePickerProps,
+  DatePickerSingleProps,
+  DatePickerRangeProps,
+} from './DatePicker';
+export type { DateRange } from 'react-day-picker';